Safari-611.1.10/LayoutTests/http/tests/security/cross-origin-cached-images-canvas-expected.txt

CONSOLE MESSAGE: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
Trying to load and render in canvas sequentially the same image from different origins (no cors mode).

PASS: Successfully retrieved image data.
PASS: The image is not same-origin with this document.
